모든엔티티, 디티오 어노테이션, 조인컬럼, 테이블명 오류 수정 , exception/jwt/Login 관련 파일 및 로직 작성, 로그인 기능 완료, 강의실 리스트 조회 기능 완료

+++ src/main/java/com/greedy/StudyFamily/lecture/entity/CourseHistory.java
@@ -34,7 +34,7 @@ public class CourseHistory {
 	private Student student;
 	
 	@Column(name = "COURSE_TIME")
-	private int courseTime;
+	private Long courseTime;
 	
 	@ManyToOne
 	@JoinColumn(name = "LECTURE_WEEK_CODE")
